pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumos-image-engine"
version = "0.1.0"
description = "Synthetic image generation with biometric non-linkability, metadata stripping, C2PA provenance, and face de-identification"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "AumOS Enterprise" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aumos-common>=0.1.0",
    "aumos-proto>=0.1.0",
    "fastapi>=0.110.0",
    "pydantic>=2.6.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "diffusers>=0.27.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "Pillow>=10.2.0",
    "c2pa-python>=0.4.0",
    "opencv-python-headless>=4.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "aiokafka>=0.10.0",
    "miniopy-async>=1.18.0",
    "structlog>=24.1.0",
    "alembic>=1.13.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "scipy>=1.12.0",
    "numpy>=1.26.0",
    "safetensors>=0.4.0",
    "peft>=0.10.0",
    "pydicom>=2.4.0",
]

[project.optional-dependencies]
benchmarks = [
    "torchmetrics[image]>=1.3.0",
    "cleanfid>=0.1.35",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "testcontainers[postgres,kafka,redis]>=3.7.0",
    "factory-boy>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aumos_image_engine"]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "B", "A", "COM", "C4", "PT", "RUF"]
ignore = ["ANN101", "ANN102", "COM812"]

[tool.ruff.lint.isort]
known-first-party = ["aumos_image_engine"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=aumos_image_engine --cov-report=term-missing --cov-fail-under=80"
